added comment to StickyAssignor
